requirements_dev.txt
pytest
pytest-mock
pytest-cov
pylint
python-telegram-bot